# services/api-gateway/docs/api-gateway-openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "TrioBank API Gateway",
        "version": "1.0.0",
        "description": "TrioBank API Gateway - Unified access point for all microservices. Handles authentication, rate limiting, and routing to Ledger, Account, and Card services.",
        "contact": {
            "name": "TrioBank Platform",
            "url": "https://triobank.example.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local Development"
        },
        {
            "url": "https://api.triobank.example.com",
            "description": "Production"
        }
    ],
...
    "paths": {
...
        "/v1/cards/debit": {
            "post": {
                "tags": [
                    "Card"
                ],
                "summary": "Issue debit card",
                "description": "Creates a new debit card linked to an account",
                "operationId": "issueDebitCard",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IssueDebitCardRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Debit card issued successfully"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "Account not found"
                    }
                }
            }
        },
        "/v1/cards/virtual": {
            "post": {
                "tags": [
                    "Card"
                ],
                "summary": "Issue virtual card",
                "description": "Creates a new virtual card for online transactions",
                "operationId": "issueVirtualCard",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IssueVirtualCardRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Virtual card issued successfully"
                    }
                }
            }
        },
        "/v1/cards": {
            "get": {
                "tags": [
                    "Card"
                ],
                "summary": "Get cards by customer or account",
                "description": "Lists cards filtered by customerId or accountId",
                "operationId": "getCards",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "accountId",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cardType",
                        "in": "query",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "enum": [
                                    "DEBIT",
                                    "VIRTUAL"
                                ]
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card list retrieved successfully"
                    }
                }
            }
        },
        "/v1/cards/{id}": {
            "get": {
                "tags": [
                    "Card"
                ],
                "summary": "Get card details",
                "description": "Retrieves details of a specific card",
                "operationId": "getCard",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card details retrieved"
                    },
                    "404": {
                        "description": "Card not found"
                    }
                }
            }
        },
        "/v1/cards/{id}/block": {
            "patch": {
                "tags": [
                    "Card"
                ],
                "summary": "Block card",
                "description": "Blocks a card with a reason (LOST, STOLEN, etc.)",
                "operationId": "blockCard",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "reason",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "LOST",
                                "STOLEN",
                                "SUSPICIOUS_ACTIVITY",
                                "USER_REQUEST"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card blocked successfully"
                    },
                    "404": {
                        "description": "Card not found"
                    }
                }
            }
        },
        "/v1/cards/{id}/activate": {
            "patch": {
                "tags": [
                    "Card"
                ],
                "summary": "Activate card",
                "description": "Activates a blocked or inactive card",
                "operationId": "activateCard",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card activated successfully"
                    },
                    "404": {
                        "description": "Card not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT token from /auth/login/confirm endpoint"
            }
        },
        "schemas": {
...
            "IssueDebitCardRequest": {
                "type": "object",
                "required": [
                    "accountId",
                    "cardholderName"
                ],
                "properties": {
                    "accountId": {
                        "type": "string"
                    },
                    "cardholderName": {
                        "type": "string",
                        "example": "AHMET YILMAZ"
                    },
                    "dailyWithdrawalLimit": {
                        "type": "number"
                    },
                    "atmEnabled": {
                        "type": "boolean"
                    },
                    "pin": {
                        "type": "string",
                        "pattern": "^[0-9]{4}$"
                    }
                }
            },
            "IssueVirtualCardRequest": {
                "type": "object",
                "required": [
                    "accountId",
                    "cardholderName"
                ],
                "properties": {
                    "accountId": {
                        "type": "string"
                    },
                    "cardholderName": {
                        "type": "string"
                    },
                    "onlineOnly": {
                        "type": "boolean",
                        "default": true
                    },
                    "singleUse": {
                        "type": "boolean",
                        "default": false
                    },
                    "singleUseValidityHours": {
                        "type": "integer",
                        "default": 24
                    },
                    "usageRestriction": {
                        "type": "string"
                    }
                }
            }
        }
    }
}